Add loadbalancer-ng apply command
With the loadbalancer-ng apply command, it is possible to use the new
playbooks to manage the load balancer. First, the load balancer
service itself is deployed. Then the configuration for all enabled
services is added in parallel.

Due to a problem with the GC of Python it is currently not possible to
execute the task completely in the background. Therefore, even if
--no-wait is specified, the task is still executed in the foreground.

Signed-off-by: Christian Berendt <berendt@osism.tech>