Add manage command
The managed command is a wrapper command for the openstack-image-manager
tool. With osism manage images you can call the openstack-image-manager,
installed in the python-osism container image.

The use is just a little bit awkward when arguments have to be given:

osism manage images -- --name Cirros

This is fixed when switching to Typer as the CLI framework.

Signed-off-by: Christian Berendt <berendt@osism.tech>