Use /etc/openstack symlink instead of OS_CLIENT_CONFIG_FILE
An explicit path to a clouds.yml must be set in OS_CLIENT_CONFIG_FILE.
This means that no secure.yml can be used. By using /etc/openstack,
one of the default locations for the configuration files of the OpenStack
SDK, the use of OS_CLIENT_CONFIG_FILE is not necessary. This way there is
no limitation that you cannot use the secure.yml.

Signed-off-by: Christian Berendt <berendt@osism.tech>