Add support for limiting selected hosts to an additional pattern (#513)
The `get hosts` command now accepts an optional `--limit` argument, which allows
users to specify an additional pattern to limit the selected hosts. This is
useful when users want to narrow down the hosts based on specific criteria. The
`limit` argument is passed to the `ansible-inventory` command as `--limit`
option.

Signed-off-by: Christian Berendt <berendt@osism.tech>