Add cold migration to osism manage compute migrate (#1287)
Automatically cold migrate instances in SHUTOFF state when issueing
`osism manage compute migrate` command.
Cold migration will be automatically confirmed if the command is waiting
for the migration to finish. A warning has been added to the description
of the `--no-wait` parameter.
A `--no-cold-migration` parameter is added to allow to skip cold
migration and thus allow to fallback to the previous behaviour.

Part of https://github.com/osism/issues/issues/1248

Signed-off-by: Jan Horstmann <horstmann@osism.tech>