Cleanup ansible requirements
The ansible collections are only required by the integrated
netbox-manager and the osism console --type ansible command. For
netbox-manager we only need the netbox collection. For the osism
console --type ansible command it's sufficient to provide basic Ansible
modules. Everything more complex should be handled in custom plays
(custom plays are executed in the osism-ansible environment).

Signed-off-by: Christian Berendt <berendt@osism.tech>